pyproject.toml
[tool.poetry]
name = "blokmap-backend"
version = "0.1.0"
description = "FastAPI implementation of the blokmap backend."
authors = [
    "Ewout Verlinde <ewout@gentsestudentenraad.be>",
    "Tibo Ulens <tibo@gentsestudentenraad.be>",
]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
gunicorn = "^23.0.0"
psycopg = "^3.2.5"
alembic = "^1.14.1"
pyjwt = "^2.10.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.20"
pydantic = "^2.10.6"
sqlmodel = "^0.0.22"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
ruff = "^0.9.7"
taskipy = "^1.14.1"

[tool.taskipy.tasks]
lint = "ruff check app/ test/ --fix"
format = "ruff format app/ test/"
test = "pytest test"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
